Move tests from test_schema into other test modules
