+++ jwst/datamodels/tests/test_multislit.py
@@ -1,0 +1,78 @@
+from astropy.io import fits
+import jsonschema
+import numpy as np
+from numpy.testing import assert_array_equal
+import pytest
+
+from jwst.datamodels import MultiSlitModel, ImageModel
+
+
+def test_multislit_model():
+    array1 = np.asarray(np.random.rand(2, 2), dtype='float32')
+    array2 = np.asarray(np.random.rand(2, 2), dtype='float32')
+
+    with MultiSlitModel() as ms:
+        assert len(ms.slits) == 0
+        ms.slits.append(ms.slits.item())
+        ms.slits[-1].data = array1
+        assert len(ms.slits) == 1
+        ms.slits.append(ms.slits.item())
+        ms.slits[-1].data = array2
+        assert len(ms.slits) == 2
+        for i, slit in enumerate(ms.slits):
+            assert slit == ms.slits[i]
+        ms2 = ms.copy()
+        assert len(ms2.slits) == 2
+        assert_array_equal(ms.slits[-1].data, array2)
+        del ms.slits[0]
+        assert len(ms.slits) == 1
+        assert_array_equal(ms.slits[0].data, array2)
+
+
+def test_copy_multslit():
+    model1 = MultiSlitModel()
+    model2 = MultiSlitModel()
+
+    model1.slits.append(ImageModel(np.ones((1024, 1024))))
+    model2.slits.append(ImageModel(np.ones((1024, 1024)) * 2))
+
+    # Create the ouput model as a copy of the first input
+    output = model1.copy()
+
+    assert len(model1.slits) == 1
+    assert len(model2.slits) == 1
+    assert len(output.slits) == 1
+
+    assert model1.slits[0].data[330, 330] == 1
+    assert output.slits[0].data[330, 330] == 1
+    assert id(model1.slits[0].data) != id(output.slits[0].data)
+
+    output.slits[0].data = model1.slits[0].data - model2.slits[0].data
+
+    assert model1.slits[0].data[330, 330] == 1
+    assert output.slits[0].data[330, 330] == -1
+
+
+def test_multislit_move_from_fits(tmp_path):
+    path = tmp_path / "multislit_move.fits"
+
+    hdulist = fits.HDUList()
+    hdulist.append(fits.PrimaryHDU())
+    for i in range(5):
+        hdu = fits.ImageHDU(data=np.zeros((64, 64)), name='SCI')
+        hdu.ver = i + 1
+        hdulist.append(hdu)
+
+    hdulist.writeto(path)
+
+    n = MultiSlitModel()
+    with MultiSlitModel(path) as m:
+        n.slits.append(m.slits[2])
+
+        assert len(n.slits) == 1
+
+
+def test_multislit_append_string():
+    with pytest.raises(jsonschema.ValidationError):
+        m = MultiSlitModel(strict_validation=True)
+        m.slits.append('junk')
